fix: show template message above step header
